requirements.txt
fastapi
uvicorn
gunicorn
pandas
numpy
openpyxl
pypdf
beautifulsoup4
openai
python-dotenv
azure-search-documents
PyJWT[crypto]